requirements.txt
Django
uwsgi
django-rest-framework
